+++ test/location_test_comment.dart
@@ -1,0 +1,2 @@
+// The LocationStreamWidget relies heavily on futures and geolocator plugin
+// and is really not suited for widget testing. It needs integration testing.